+++ wordless/wl_settings/wl_settings_lemmatization.py
@@ -33,8 +33,8 @@ class Wl_Worker_Preview_Lemmatizer(wl_threading.Wl_Worker_No_Progress):
     def run(self):
         preview_results = []
 
-        preview_lang = self.main.settings_custom['lemmatization']['preview_lang']
-        preview_samples = self.main.settings_custom['lemmatization']['preview_samples']
+        preview_lang = self.main.settings_custom['lemmatization']['preview']['preview_lang']
+        preview_samples = self.main.settings_custom['lemmatization']['preview']['preview_samples']
 
         for line in preview_samples.split('\n'):
             line = line.strip()
@@ -141,18 +141,18 @@ class Wl_Settings_Lemmatization(wl_settings.Wl_Settings_Node):
         self.layout().setRowStretch(1, 2)
 
     def lemmatizers_changed(self, lang):
-        if lang == self.settings_custom['preview_lang']:
+        if lang == self.settings_custom['preview']['preview_lang']:
             self.preview_results_changed()
 
     def preview_changed(self):
-        self.settings_custom['preview_lang'] = wl_conversion.to_lang_code(self.main, self.combo_box_lemmatization_preview_lang.currentText())
-        self.settings_custom['preview_samples'] = self.text_edit_lemmatization_preview_samples.toPlainText()
-        self.settings_custom['preview_results'] = self.text_edit_lemmatization_preview_results.toPlainText()
+        self.settings_custom['preview']['preview_lang'] = wl_conversion.to_lang_code(self.main, self.combo_box_lemmatization_preview_lang.currentText())
+        self.settings_custom['preview']['preview_samples'] = self.text_edit_lemmatization_preview_samples.toPlainText()
+        self.settings_custom['preview']['preview_results'] = self.text_edit_lemmatization_preview_results.toPlainText()
 
     def preview_results_changed(self):
-        if self.settings_custom['preview_samples']:
+        if self.settings_custom['preview']['preview_samples']:
             if self.combo_box_lemmatization_preview_lang.isEnabled():
-                row = list(self.settings_global.keys()).index(self.settings_custom['preview_lang'])
+                row = list(self.settings_global.keys()).index(self.settings_custom['preview']['preview_lang'])
 
                 self.table_lemmatizers.itemDelegateForRow(row).set_enabled(False)
                 self.combo_box_lemmatization_preview_lang.setEnabled(False)
@@ -182,7 +182,7 @@ class Wl_Settings_Lemmatization(wl_settings.Wl_Settings_Node):
         self.button_lemmatization_show_preview.setText(self.tr('Show preview'))
         self.text_edit_lemmatization_preview_results.setPlainText('\n'.join(preview_results))
 
-        row = list(self.settings_global.keys()).index(self.settings_custom['preview_lang'])
+        row = list(self.settings_global.keys()).index(self.settings_custom['preview']['preview_lang'])
 
         self.table_lemmatizers.itemDelegateForRow(row).set_enabled(True)
         self.combo_box_lemmatization_preview_lang.setEnabled(True)
@@ -197,11 +197,11 @@ class Wl_Settings_Lemmatization(wl_settings.Wl_Settings_Node):
 
         self.table_lemmatizers.disable_updates()
 
-        for i, lang in enumerate(settings['lemmatizers']):
+        for i, lang in enumerate(settings['lemmatizer_settings']):
             self.table_lemmatizers.model().item(i, 1).setText(wl_nlp_utils.to_lang_util_text(
                 self.main,
                 util_type = 'lemmatizers',
-                util_code = settings['lemmatizers'][lang]
+                util_code = settings['lemmatizer_settings'][lang]
             ))
 
         self.table_lemmatizers.enable_updates()
@@ -210,16 +210,16 @@ class Wl_Settings_Lemmatization(wl_settings.Wl_Settings_Node):
             self.combo_box_lemmatization_preview_lang.blockSignals(True)
             self.text_edit_lemmatization_preview_samples.blockSignals(True)
 
-            self.combo_box_lemmatization_preview_lang.setCurrentText(wl_conversion.to_lang_text(self.main, settings['preview_lang']))
-            self.text_edit_lemmatization_preview_samples.setText(settings['preview_samples'])
-            self.text_edit_lemmatization_preview_results.setText(settings['preview_results'])
+            self.combo_box_lemmatization_preview_lang.setCurrentText(wl_conversion.to_lang_text(self.main, settings['preview']['preview_lang']))
+            self.text_edit_lemmatization_preview_samples.setText(settings['preview']['preview_samples'])
+            self.text_edit_lemmatization_preview_results.setText(settings['preview']['preview_results'])
 
             self.combo_box_lemmatization_preview_lang.blockSignals(False)
             self.text_edit_lemmatization_preview_samples.blockSignals(False)
 
     def apply_settings(self):
-        for i, lang in enumerate(self.settings_custom['lemmatizers']):
-            self.settings_custom['lemmatizers'][lang] = wl_nlp_utils.to_lang_util_code(
+        for i, lang in enumerate(self.settings_custom['lemmatizer_settings']):
+            self.settings_custom['lemmatizer_settings'][lang] = wl_nlp_utils.to_lang_util_code(
                 self.main,
                 util_type = 'lemmatizers',
                 util_text = self.table_lemmatizers.model().item(i, 1).text()